Improve the BaseUrl middleware
The old BaseUrl middleware was very naive, it simply set
the url field of the connection to whatever was provided.
This new version actually parses the URL and sets the different
Conn fields accordingly. It also validates that the base url actually
contains a host, so that a valid request can be made.